remove board api. reduce outlet api to only support get list. load outlets from config straight
