+++ controller/controller.go
@@ -34,7 +34,6 @@ func New(config Config) (*Controller, error) {
 
 func (c *Controller) createBuckets() error {
 	buckets := []string{
-		BoardBucket,
 		EquipmentBucket,
 		JobBucket,
 		OutletBucket,
