+++ jsx/ui.jsx
@@ -2,7 +2,6 @@ import React from 'react'
 import { render } from 'react-dom'
 import { Tab, Tabs, TabList, TabPanel } from 'react-tabs'
 import Dashboard from './dashboard.jsx'
-import Electronics from './electronics.jsx'
 import Equipments from './equipments.jsx'
 import Jobs from './jobs.jsx'
 import ATO from './ato.jsx'
@@ -17,7 +16,6 @@ export default class App extends React.Component {
       <Tabs onSelect={this.handleSelect} selectedIndex={0}>
         <TabList>
           <Tab>Dashboard</Tab>
-          <Tab>Electronics</Tab>
           <Tab>Equipments</Tab>
           <Tab>Jobs</Tab>
           <Tab>ATO</Tab>
@@ -25,9 +23,6 @@ export default class App extends React.Component {
         </TabList>
         <TabPanel>
           <Dashboard />
-        </TabPanel>
-        <TabPanel>
-          <Electronics />
         </TabPanel>
         <TabPanel>
           <Equipments />
